Fixing stuff

+++ train_network.py
@@ -983,7 +983,7 @@ class NetworkTrainer:
                         else:
                             target = noise
                         if args.reg_image_training:
-                            loss = torch.nn.functional.mse_loss(noisy_latents.float() - noise_pred.float(), cond_latents, reduction="none")
+                            loss = torch.nn.functional.mse_loss(noisy_latents.float() - noise_pred.float(), cond_latents.float(), reduction="none")
                         else:
                             loss = torch.nn.functional.mse_loss(noise_pred.float(), target.float(), reduction="none")
